feat(www/audio): extract item from LatestArticles list into component and add loading state
